# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "FOODIO API",
    "description": "FOODIO API Swagger documentation for the food delivery platform.",
    "version": "1.0.0",
    "contact": {
      "name": "FOODIO API Team",
      "email": "support@foodio.example"
    }
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Primary server"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/chatbot/message": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "example": "any"
                },
                "conversationId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ]
      }
    },
    "/api/chatbot/conversation/{conversationId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ]
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ]
      }
    },
    "/api/chatbot/conversations": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ]
      }
    },
    "/api/chatbot/recommendations": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cuisineType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "budget",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dietary",
            "in": "query",
            "type": "string"
          },
          {
            "name": "mood",
            "in": "query",
            "type": "string"
          },
          {
            "name": "deliveryTime",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ]
      }
    },
    "/api/chatbot/smart-recommendation": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ]
      }
    },
    "/api/chatbot/stats": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ]
      }
    },
    "/api/chatbot/feedback": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "messageId": {
                  "example": "any"
                },
                "rating": {
                  "example": "any"
                },
                "feedbackText": {
                  "example": "any"
                },
                "feedbackType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ]
      }
    },
...
  }
}